fixes #10005

make sure that the candidate for new container network does not overlap with any existing local networks

**notes**:
- the issue seen below with stale/inactive kvm network (even after `--delete` flag) is resolved in pr #10479 
- we can try to use new `GetNetworkFrom()` (in pkg/network/network.go) for getting any interface's `mtu` (i know there were some issues with it)
- having params in kvm's `networkTmpl` opens possibility to have different private network for each cluster (needs more work to have that if we want to); ref: https://github.com/kubernetes/minikube/issues/9049#issuecomment-679192676


## example


### before

set docker net to kvm's default (static) one to simulate conflict:
```
❯ grep "const firstSubnetAddr" pkg/drivers/kic/oci/network_create.go
const firstSubnetAddr = "192.168.39.0"

❯ grep "ip address" pkg/drivers/kvm/network.go
  <ip address='192.168.39.1' netmask='255.255.255.0'>
```

```
❯ out/minikube start --driver=docker -p docker-192.168.39.0
😄  [docker-192.168.39.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the docker driver based on user configuration
❗  docker is currently using the btrfs storage driver, consider switching to overlay2 for better performance
👍  Starting control plane node docker-192.168.39.0 in cluster docker-192.168.39.0
🔥  Creating docker container (CPUs=2, Memory=16000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "docker-192.168.39.0" cluster and "default" namespace by default
```

```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           DOWN           192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
br-08ada8d5dfa4  UP             172.22.0.1/16
docker0          DOWN           172.17.0.1/16
br-ec7e0d3126e7  UP             192.168.39.1/24
```

```
❯ out/minikube start --driver=kvm -p kvm-192.168.39.0
😄  [kvm-192.168.39.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the kvm2 driver based on user configuration
💿  Downloading VM boot image ...
    > minikube-v1.17.0.iso.sha256: 65 B / 65 B [-------------] 100.00% ? p/s 0s
    > minikube-v1.17.0.iso: 212.69 MiB / 212.69 MiB [] 100.00% 24.09 MiB p/s 9s
👍  Starting control plane node kvm-192.168.39.0 in cluster kvm-192.168.39.0
💾  Downloading Kubernetes v1.20.2 preload ...
    > preloaded-images-k8s-v8-v1....: 491.22 MiB / 491.22 MiB  100.00% 21.47 Mi
🔥  Creating kvm2 VM (CPUs=2, Memory=6000MB, Disk=20000MB) ...
🤦  StartHost failed, but will try again: creating host: create: Error creating machine: Error in driver during machine creation: creating network: creating network minikube-net: virError(Code=1, Domain=0, Message='internal error: Network is already in use by interface br-ec7e0d3126e7')
🔄  Restarting existing kvm2 VM for "kvm-192.168.39.0" ...
😿  Failed to start kvm2 VM. Running "minikube delete -p kvm-192.168.39.0" may fix it: driver start: ensuring active networks: deleting inoperable network minikube-net: destroying network: virError(Code=55, Domain=19, Message='Requested operation is not valid: network 'minikube-net' is not active')

❌  Exiting due to GUEST_PROVISION: Failed to start host: driver start: ensuring active networks: deleting inoperable network minikube-net: destroying network: virError(Code=55, Domain=19, Message='Requested operation is not valid: network 'minikube-net' is not active')

😿  If the above advice does not help, please let us know:
👉  https://github.com/kubernetes/minikube/issues/new/choose
```

```
❯ minikube profile list
|---------------------|-----------|---------|--------------|------|---------|---------|-------|
|       Profile       | VM Driver | Runtime |      IP      | Port | Version | Status  | Nodes |
|---------------------|-----------|---------|--------------|------|---------|---------|-------|
| docker-192.168.39.0 | docker    | docker  | 192.168.39.2 | 8443 | v1.20.2 | Running |     1 |
| kvm-192.168.39.0    | kvm2      | docker  |              | 8443 | v1.20.2 | Unknown |     1 |
|---------------------|-----------|---------|--------------|------|---------|---------|-------|
```

```
❯ minikube delete --all
🔥  Deleting "docker-192.168.39.0" in docker ...
🔥  Removing /home/prezha/.minikube/machines/docker-192.168.39.0 ...
💀  Removed all traces of the "docker-192.168.39.0" cluster.
🔥  Deleting "kvm-192.168.39.0" in kvm2 ...
💀  Removed all traces of the "kvm-192.168.39.0" cluster.
🔥  Successfully deleted all profiles
```

```
# virsh net-list --all
 Name           State      Autostart   Persistent
---------------------------------------------------
 default        active     yes         yes
 minikube-net   inactive   yes         yes
```

### after

simulate conflict:
```
❯ grep "const firstSubnetAddr" pkg/drivers/kic/oci/network_create.go
const firstSubnetAddr = "192.168.49.0"

❯ grep "const firstSubnetAddr" pkg/drivers/kvm/network.go
const firstSubnetAddr = "192.168.49.0"
```

```
❯ out/minikube start --driver=docker -p docker-192.168.49.0
😄  [docker-192.168.49.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the docker driver based on user configuration
❗  docker is currently using the btrfs storage driver, consider switching to overlay2 for better performance
👍  Starting control plane node docker-192.168.49.0 in cluster docker-192.168.49.0
🔥  Creating docker container (CPUs=2, Memory=16000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "docker-192.168.49.0" cluster and "default" namespace by default
```

```
❯ out/minikube start --driver=kvm -p kvm-192.168.49.0
😄  [kvm-192.168.49.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the kvm2 driver based on user configuration
👍  Starting control plane node kvm-192.168.49.0 in cluster kvm-192.168.49.0
💾  Downloading Kubernetes v1.20.2 preload ...
    > preloaded-images-k8s-v8-v1....: 491.22 MiB / 491.22 MiB  100.00% 25.11 Mi
🔥  Creating kvm2 VM (CPUs=2, Memory=6000MB, Disk=20000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "kvm-192.168.49.0" cluster and "default" namespace by default
```

```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           UP             192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
br-08ada8d5dfa4  UP             172.22.0.1/16
docker0          DOWN           172.17.0.1/16
br-33a6261ee1e8  UP             192.168.49.1/24
virbr1           UP             192.168.69.1/24
```

```
❯ minikube profile list
|---------------------|-----------|---------|---------------|------|---------|---------|-------|
|       Profile       | VM Driver | Runtime |      IP       | Port | Version | Status  | Nodes |
|---------------------|-----------|---------|---------------|------|---------|---------|-------|
| docker-192.168.49.0 | docker    | docker  | 192.168.49.2  | 8443 | v1.20.2 | Running |     1 |
| kvm-192.168.49.0    | kvm2      | docker  | 192.168.69.72 | 8443 | v1.20.2 | Running |     1 |
|---------------------|-----------|---------|---------------|------|---------|---------|-------|
```

```
❯ minikube delete --all
🔥  Deleting "docker-192.168.49.0" in docker ...
🔥  Removing /home/prezha/.minikube/machines/docker-192.168.49.0 ...
💀  Removed all traces of the "docker-192.168.49.0" cluster.
🔥  Deleting "kvm-192.168.49.0" in kvm2 ...
💀  Removed all traces of the "kvm-192.168.49.0" cluster.
🔥  Successfully deleted all profiles
```

```
codex:~ # virsh net-list --all
 Name      State    Autostart   Persistent
--------------------------------------------
 default   active   yes         yes
```
---

same goes for the other side:

```
❯ grep "const firstSubnetAddr" pkg/drivers/kic/oci/network_create.go
const firstSubnetAddr = "192.168.39.0"

❯ grep "const firstSubnetAddr" pkg/drivers/kvm/network.go
const firstSubnetAddr = "192.168.39.0"
```

```
❯ out/minikube start --driver=kvm -p kvm-192.168.39.0
😄  [kvm-192.168.39.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the kvm2 driver based on user configuration
👍  Starting control plane node kvm-192.168.39.0 in cluster kvm-192.168.39.0
🔥  Creating kvm2 VM (CPUs=2, Memory=6000MB, Disk=20000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "kvm-192.168.39.0" cluster and "default" namespace by default
```

```
❯ out/minikube start --driver=docker -p docker-192.168.39.0
😄  [docker-192.168.39.0] minikube v1.17.1 on Opensuse-Tumbleweed
✨  Using the docker driver based on user configuration
❗  docker is currently using the btrfs storage driver, consider switching to overlay2 for better performance
👍  Starting control plane node docker-192.168.39.0 in cluster docker-192.168.39.0
🔥  Creating docker container (CPUs=2, Memory=16000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "docker-192.168.39.0" cluster and "default" namespace by default
```

```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           UP             192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
br-08ada8d5dfa4  UP             172.22.0.1/16
docker0          DOWN           172.17.0.1/16
virbr1           UP             192.168.39.1/24
br-8616cbd0a692  UP             192.168.49.1/24
```

```
❯ minikube profile list
|---------------------|-----------|---------|----------------|------|---------|---------|-------|
|       Profile       | VM Driver | Runtime |       IP       | Port | Version | Status  | Nodes |
|---------------------|-----------|---------|----------------|------|---------|---------|-------|
| docker-192.168.39.0 | docker    | docker  | 192.168.49.2   | 8443 | v1.20.2 | Running |     1 |
| kvm-192.168.39.0    | kvm2      | docker  | 192.168.39.174 | 8443 | v1.20.2 | Running |     1 |
|---------------------|-----------|---------|----------------|------|---------|---------|-------|
```
---
## additional examples (as per @medyagh's suggestion)

### current master

```
❯ docker network list
NETWORK ID     NAME                 DRIVER    SCOPE
d4c6d1af63d9   br-test00            bridge    local
4caf32131f3f   br-test01            bridge    local
bafaa95e04c9   br-test02            bridge    local
82697145172a   br-test03            bridge    local
3fb60e8289eb   br-test04            bridge    local
f6bd05bf4625   br-test05            bridge    local
f62e9c91fe42   br-test06            bridge    local
a4dc3444d406   br-test07            bridge    local
661f7a0277f9   br-test08            bridge    local
de4b38920f04   br-test09            bridge    local
9816f4fd95a3   bridge               bridge    local
790f8998f0f6   host                 host      local
a9347ae22f3e   minikube             bridge    local
ee746e6d9be7   none                 null      local
```
```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           DOWN           192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
docker0          DOWN           172.17.0.1/16
br-d4c6d1af63d9  DOWN           192.168.49.1/24
br-4caf32131f3f  DOWN           192.168.59.1/24
br-bafaa95e04c9  DOWN           192.168.70.1/24
br-82697145172a  DOWN           192.168.82.1/24
br-3fb60e8289eb  DOWN           192.168.95.1/24
br-f6bd05bf4625  DOWN           192.168.109.1/24
br-f62e9c91fe42  DOWN           192.168.124.1/24
br-a4dc3444d406  DOWN           192.168.140.1/24
br-661f7a0277f9  DOWN           192.168.157.1/24
br-de4b38920f04  DOWN           192.168.175.1/24
```
```
❯ minikube start                             
stacklog: logging to stack.slog, sampling every 125ms
😄  minikube v1.17.1 on Opensuse-Tumbleweed 
✨  Automatically selected the docker driver. Other choices: kvm2, ssh
❗  docker is currently using the btrfs storage driver, consider switching to overlay2 for better performance
👍  Starting control plane node minikube in cluster minikube
🔥  Creating docker container (CPUs=2, Memory=16000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "minikube" cluster and "default" namespace by default
stacklog: stopped. stored 428 samples to stack.slog
```
```
 ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           DOWN           192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
docker0          DOWN           172.17.0.1/16
br-d4c6d1af63d9  DOWN           192.168.49.1/24
br-4caf32131f3f  DOWN           192.168.59.1/24
br-bafaa95e04c9  DOWN           192.168.70.1/24
br-82697145172a  DOWN           192.168.82.1/24
br-3fb60e8289eb  DOWN           192.168.95.1/24
br-f6bd05bf4625  DOWN           192.168.109.1/24
br-f62e9c91fe42  DOWN           192.168.124.1/24
br-a4dc3444d406  DOWN           192.168.140.1/24
br-661f7a0277f9  DOWN           192.168.157.1/24
br-de4b38920f04  DOWN           192.168.175.1/24
br-a9347ae22f3e  UP             192.168.194.1/24
```

`slowjam` [oci.CreateNetwork] duration: 1 sec
> [ '265: machine.timedCreateHost', 'oci.CreateNetwork', '#808000',  new Date(1874), new Date(2874) ],

### pull request

```
❯ docker network list
NETWORK ID     NAME                 DRIVER    SCOPE
46c634fb4fda   br-test00            bridge    local
2c6a49c30e38   br-test01            bridge    local
6c6ad70ee603   br-test02            bridge    local
83e070309643   br-test03            bridge    local
2cac664e256c   br-test04            bridge    local
8fd544819864   br-test05            bridge    local
bacded214b6b   br-test06            bridge    local
40d9ff702bdc   br-test07            bridge    local
57f7eb9a4178   br-test08            bridge    local
1c6e55082f96   br-test09            bridge    local
9816f4fd95a3   bridge               bridge    local
790f8998f0f6   host                 host      local
817b88552774   minikube             bridge    local
ee746e6d9be7   none                 null      local
```
```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           DOWN           192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
br-08ada8d5dfa4  UP             172.22.0.1/16
docker0          DOWN           172.17.0.1/16
br-46c634fb4fda  DOWN           192.168.49.1/24
br-2c6a49c30e38  DOWN           192.168.59.1/24
br-6c6ad70ee603  DOWN           192.168.69.1/24
br-83e070309643  DOWN           192.168.79.1/24
br-2cac664e256c  DOWN           192.168.89.1/24
br-8fd544819864  DOWN           192.168.99.1/24
br-bacded214b6b  DOWN           192.168.109.1/24
br-40d9ff702bdc  DOWN           192.168.119.1/24
br-57f7eb9a4178  DOWN           192.168.129.1/24
br-1c6e55082f96  DOWN           192.168.139.1/24
```
```
❯ minikube start
stacklog: logging to stack.slog, sampling every 125ms
😄  minikube v1.17.1 on Opensuse-Tumbleweed
✨  Automatically selected the docker driver. Other choices: kvm2, ssh
❗  docker is currently using the btrfs storage driver, consider switching to overlay2 for better performance
👍  Starting control plane node minikube in cluster minikube
🔥  Creating docker container (CPUs=2, Memory=16000MB) ...
🐳  Preparing Kubernetes v1.20.2 on Docker 20.10.2 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v4
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "minikube" cluster and "default" namespace by default
stacklog: stopped. stored 422 samples to stack.slog
```
```
❯ ip -4 -br -o a s
lo               UNKNOWN        127.0.0.1/8
virbr0           DOWN           192.168.122.1/24
wlp113s0         UP             192.168.42.17/24
docker0          DOWN           172.17.0.1/16
br-46c634fb4fda  DOWN           192.168.49.1/24
br-2c6a49c30e38  DOWN           192.168.59.1/24
br-6c6ad70ee603  DOWN           192.168.69.1/24
br-83e070309643  DOWN           192.168.79.1/24
br-2cac664e256c  DOWN           192.168.89.1/24
br-8fd544819864  DOWN           192.168.99.1/24
br-bacded214b6b  DOWN           192.168.109.1/24
br-40d9ff702bdc  DOWN           192.168.119.1/24
br-57f7eb9a4178  DOWN           192.168.129.1/24
br-1c6e55082f96  DOWN           192.168.139.1/24
br-817b88552774  UP             192.168.149.1/24
```

`slowjam` [oci.CreateNetwork] duration: 0.751 sec
> [ '278: machine.timedCreateHost', 'oci.CreateNetwork', '#da70d6',  new Date(1749), new Date(2500) ],